+++ legadilo/feeds/management/commands/clean_data.py
@@ -17,10 +17,11 @@
 from django.core.management import BaseCommand
 
 from legadilo.feeds.models import Feed
-from legadilo.reading.models import ArticleFetchError
+from legadilo.reading.models import Article, ArticleFetchError
 
 
 class Command(BaseCommand):
     def handle(self, *args, **options):
         Feed.objects.get_feed_update_for_cleanup().delete()
         ArticleFetchError.objects.get_queryset().for_cleanup().delete()
+        Article.objects.get_queryset().for_cleanup().delete()