+++ legadilo/reading/models/article.py
@@ -38,7 +38,11 @@ from legadilo.utils.collections_utils import max_or_none, min_or_none
 from legadilo.utils.security import full_sanitize
 from legadilo.utils.text import get_nb_words_from_html
 from legadilo.utils.time_utils import utcnow
-from legadilo.utils.validators import language_code_validator, list_of_strings_json_schema_validator
+from legadilo.utils.validators import (
+    language_code_validator,
+    list_of_strings_json_schema_validator,
+    table_of_content_json_schema_validator,
+)
 
 from .article_fetch_error import ArticleFetchError
 
@@ -362,6 +366,36 @@ class ArticleQuerySet(models.QuerySet["Article"]):
             alias_date_field_order=Coalesce(models.F("updated_at"), models.F("published_at"))
         ).order_by(order)
 
+    def for_cleanup(self) -> Self:
+        return self.alias(
+            min_feed_retention_time=models.Min("feeds__article_retention_time"),
+            # Let's keep the article until the last feed says it's time to collect.
+            feed_retention_time=models.Max("feeds__article_retention_time"),
+            article_cleanup_time=models.ExpressionWrapper(
+                models.F("read_at__epoch")
+                + (
+                    # This time is stored in days in the database. Convert it to seconds to add
+                    # it to the read at time stamp.
+                    models.F("feed_retention_time")
+                    * models.Value(
+                        24,
+                        output_field=models.IntegerField(),
+                    )
+                    * models.Value(60, output_field=models.IntegerField())
+                    * models.Value(60, output_field=models.IntegerField())
+                ),
+                output_field=models.IntegerField(),
+            ),
+        ).filter(
+            is_read=True,
+            feed_retention_time__isnull=False,
+            feed_retention_time__gt=0,
+            # If the min retention time is 0, it means at least one feed requires us to keep
+            # the article forever.
+            min_feed_retention_time__gt=0,
+            article_cleanup_time__lt=utcnow().timestamp(),
+        )
+
     def for_search(self, query: SearchQuery) -> Self:
         return (
             self.alias(search=SEARCH_VECTOR, rank=SearchRank(SEARCH_VECTOR, query))
@@ -427,6 +461,7 @@ class ArticleManager(models.Manager["Article"]):
                         slug=slugify(article_data.title),
                         summary=article_data.summary,
                         content=article_data.content,
+                        table_of_content=article_data.table_of_content,
                         reading_time=get_nb_words_from_html(article_data.content)
                         // user.settings.default_reading_time,
                         authors=article_data.authors,
@@ -643,6 +678,12 @@ class Article(models.Model):
         help_text=_("The language code for this article"),
         validators=[language_code_validator],
     )
+    table_of_content = models.JSONField(
+        validators=[table_of_content_json_schema_validator],
+        blank=True,
+        default=list,
+        help_text=_("The table of content of the article."),
+    )
 
     read_at = models.DateTimeField(null=True, blank=True)
     is_read = models.GeneratedField(
@@ -745,6 +786,7 @@ class Article(models.Model):
             )
             self.summary = article_data.summary or self.summary
             self.content = article_data.content or self.content
+            self.table_of_content = article_data.table_of_content or self.table_of_content
             self.reading_time = (
                 get_nb_words_from_html(self.content) // self.user.settings.default_reading_time
             ) or self.reading_time
@@ -757,6 +799,7 @@ class Article(models.Model):
             self.published_at = min_or_none([article_data.published_at, self.published_at])
         elif has_content_unlike_saved:
             self.content = article_data.content
+            self.table_of_content = article_data.table_of_content
 
         self.article_to_update = utcnow()