+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Changelog
 
 
+## 24.09.1
+
+### Breaking change
+
+- Command `cleanup_old_updates` has been renamed into `clean_data`.
+  - It still cleans up old feed updates and articles fetch errors.
+  - It will also clean old articles from feeds if the _Keep articles_ option is not set to _Always_ in the configuration of the feeds.
+
+### Other changes
+
+- Display a count of unread notifications.
+  - This makes the link more visible when a user has unread notifications.
+- Correct text breaks of summary.
+- Re-enable feed if we re-add a disabled feed.
+- Correct article author in page metadata.
+- Display a navigable table of content on the side of article details.
+
+
 ## 24.08.6
 
 - Can disable read on scroll temporarily.